♻️ refactor QASM import functionality and remove deprecated formats
- ♻️ turn QASM parser into its own library
- 🔥 remove import functionality from `QuantumComputation`
  the respective functionality is now provided by a separate `MQT::CoreQASM` library
- ♻️ refactor QASM parser and improve imports
