+++ include/mqt-core/qasm3/NestedEnvironment.hpp
