+++ include/mqt-core/qasm3/Statement.hpp
@@ -9,12 +9,10 @@
 
 #pragma once
 
-#include "InstVisitor.hpp"
-#include "Types.hpp"
+#include "Statement_fwd.hpp" // IWYU pragma: export
+#include "Types_fwd.hpp"
 #include "ir/Permutation.hpp"
-#include "ir/operations/ClassicControlledOperation.hpp"
 
-#include <cassert>
 #include <cstddef>
 #include <cstdint>
 #include <memory>
@@ -24,7 +22,13 @@
 #include <variant>
 #include <vector>
 
+namespace qc {
+// forward declarations
+enum ComparisonKind : std::uint8_t;
+} // namespace qc
+
 namespace qasm3 {
+class InstVisitor;
 
 struct DebugInfo {
   size_t line;
@@ -57,7 +61,7 @@ public:
   explicit DeclarationExpression(std::shared_ptr<Expression> expr)
       : expression(std::move(expr)) {}
 
-  virtual ~DeclarationExpression() = default;
+  ~DeclarationExpression() = default;
 };
 
 class Constant final : public Expression {
@@ -246,18 +250,9 @@ public:
                            std::shared_ptr<IdentifierList> params,
                            std::shared_ptr<IdentifierList> qbits,
                            std::vector<std::shared_ptr<QuantumStatement>> stmts,
-                           const bool opaque = false)
-      : Statement(std::move(debug)), identifier(std::move(id)),
-        parameters(std::move(params)), qubits(std::move(qbits)),
-        statements(std::move(stmts)), isOpaque(opaque) {
-    if (opaque) {
-      assert(statements.empty() && "Opaque gate should not have statements.");
-    }
-  }
+                           const bool opaque = false);
 
-  void accept(InstVisitor* visitor) override {
-    visitor->visitGateStatement(shared_from_this());
-  }
+  void accept(InstVisitor* visitor) override;
 };
 
 class VersionDeclaration final
@@ -270,9 +265,7 @@ public:
                               const double versionNum)
       : Statement(std::move(debug)), version(versionNum) {}
 
-  void accept(InstVisitor* visitor) override {
-    visitor->visitVersionDeclaration(shared_from_this());
-  }
+  void accept(InstVisitor* visitor) override;
 };
 
 class InitialLayout final : public Statement,
@@ -284,9 +277,7 @@ public:
       : Statement(std::move(debug)), permutation(std::move(perm)) {}
 
 private:
-  void accept(InstVisitor* visitor) override {
-    visitor->visitInitialLayout(shared_from_this());
-  }
+  void accept(InstVisitor* visitor) override;
 };
 
 class OutputPermutation final
@@ -300,9 +291,7 @@ public:
       : Statement(std::move(debug)), permutation(std::move(perm)) {}
 
 private:
-  void accept(InstVisitor* visitor) override {
-    visitor->visitOutputPermutation(shared_from_this());
-  }
+  void accept(InstVisitor* visitor) override;
 };
 
 class DeclarationStatement final
@@ -320,9 +309,7 @@ public:
       : Statement(std::move(debug)), isConst(declIsConst), type(ty),
         identifier(std::move(id)), expression(std::move(expr)) {}
 
-  void accept(InstVisitor* visitor) override {
-    visitor->visitDeclarationStatement(shared_from_this());
-  }
+  void accept(InstVisitor* visitor) override;
 };
 
 class GateModifier : public std::enable_shared_from_this<GateModifier> {
@@ -372,9 +359,7 @@ public:
         modifiers(std::move(modifierList)), arguments(std::move(argumentList)),
         operands(std::move(operandList)) {}
 
-  void accept(InstVisitor* visitor) override {
-    visitor->visitGateCallStatement(shared_from_this());
-  }
+  void accept(InstVisitor* visitor) override;
 };
 
 class AssignmentStatement final
@@ -407,9 +392,7 @@ public:
       : Statement(std::move(debug)), type(ty), identifier(std::move(id)),
         indexExpression(std::move(indexExpr)), expression(std::move(expr)) {}
 
-  void accept(InstVisitor* visitor) override {
-    visitor->visitAssignmentStatement(shared_from_this());
-  }
+  void accept(InstVisitor* visitor) override;
 };
 
 class BarrierStatement final
@@ -422,9 +405,7 @@ public:
                             std::vector<std::shared_ptr<GateOperand>> gateList)
       : QuantumStatement(std::move(debug)), gates(std::move(gateList)) {}
 
-  void accept(InstVisitor* visitor) override {
-    visitor->visitBarrierStatement(shared_from_this());
-  }
+  void accept(InstVisitor* visitor) override;
 };
 
 class ResetStatement final
@@ -437,9 +418,7 @@ public:
                           std::shared_ptr<GateOperand> g)
       : QuantumStatement(std::move(debug)), gate(std::move(g)) {}
 
-  void accept(InstVisitor* visitor) override {
-    visitor->visitResetStatement(shared_from_this());
-  }
+  void accept(InstVisitor* visitor) override;
 };
 
 class IfStatement final : public Statement,
@@ -456,8 +435,6 @@ public:
       : Statement(std::move(debug)), condition(cond), thenStatements(thenStmts),
         elseStatements(elseStmts) {}
 
-  void accept(InstVisitor* visitor) override {
-    visitor->visitIfStatement(shared_from_this());
-  }
+  void accept(InstVisitor* visitor) override;
 };
 } // namespace qasm3
