+++ include/mqt-core/qasm3/Statement_fwd.hpp
@@ -1,0 +1,36 @@
+/*
+ * Copyright (c) 2025 Chair for Design Automation, TUM
+ * All rights reserved.
+ *
+ * SPDX-License-Identifier: MIT
+ *
+ * Licensed under the MIT License
+ */
+
+#pragma once
+
+namespace qasm3 {
+struct DebugInfo;
+class Statement;
+class IfStatement;
+class IdentifierList;
+class MeasureExpression;
+class GateCallStatement;
+class GateDeclaration;
+class BarrierStatement;
+class ResetStatement;
+class AssignmentStatement;
+class VersionDeclaration;
+class DeclarationExpression;
+class DeclarationStatement;
+class GateOperand;
+class GateModifier;
+class QuantumStatement;
+class InitialLayout;
+class OutputPermutation;
+class Expression;
+class Constant;
+class BinaryExpression;
+class UnaryExpression;
+class IdentifierExpression;
+} // namespace qasm3
