+++ include/mqt-core/qasm3/Types_fwd.hpp
@@ -1,0 +1,21 @@
+/*
+ * Copyright (c) 2025 Chair for Design Automation, TUM
+ * All rights reserved.
+ *
+ * SPDX-License-Identifier: MIT
+ *
+ * Licensed under the MIT License
+ */
+
+#pragma once
+
+#include <cstdint>
+#include <memory>
+
+namespace qasm3 {
+class Expression;
+
+template <typename T> class Type;
+using TypeExpr = Type<std::shared_ptr<Expression>>;
+using ResolvedType = Type<uint64_t>;
+} // namespace qasm3
