+++ src/CMakeLists.txt
@@ -69,6 +69,9 @@ endif()
 # add the IR package
 add_subdirectory(ir)
 
+# add the QASM package
+add_subdirectory(qasm3)
+
 # add the algorithms package
 add_subdirectory(algorithms)
