+++ test/circuit_optimizer/CMakeLists.txt
@@ -9,5 +9,5 @@ if(TARGET MQT::CoreCircuitOptimizer)
   file(GLOB_RECURSE CIRCUIT_OPTIMIZER_TEST_SOURCES *.cpp)
   package_add_test(mqt-core-circuit-optimizer-test MQT::CoreCircuitOptimizer
                    ${CIRCUIT_OPTIMIZER_TEST_SOURCES})
-  target_link_libraries(mqt-core-circuit-optimizer-test PRIVATE MQT::CoreAlgorithms)
+  target_link_libraries(mqt-core-circuit-optimizer-test PRIVATE MQT::CoreAlgorithms MQT::CoreQASM)
 endif()
