--- test/circuits/test.real
@@ -1,42 +1,0 @@
-# declare number of variables
-.numvars 4
-# declare variable names
-.variables a b c d
-.constants --01
-.garbage --11
-
-# begin circuit description
-.begin
-01 a
-id1 a
-h1 a
-n1 a
-c a b
-x1 a
-y1 a
-s1 a
-si1 a
-sp1 a
-s+1 a
-v a b
-vp a b
-v+ a b
-rx1:2 a
-ry1:2 a
-rz1:2 a
-f a b
-if a b
-p a b c
-pi a b c
-p+ a b c
-q1:1 a
-q1:2 a
-q1:-2 a
-q1:4 a
-q1:-4 a
-q1:8 a
-q1:-8 a
-t2 a b
-t3 a b c
-t4 a b c d
-.end
