--- test/circuits/test.tfc
@@ -1,50 +1,0 @@
-# Example .tfc format:
-#
-#   Tags:
-#   (Note: There must be whitespace of some sort between the tag and
-#   parameters (with the exception of the comment tag, #)
-#
-#   	#  - Indicates comment.  Any text following a # on the beginning
-#            of a line will be ignored.
-#   	.v - Variable list.  All variables (in AND out) need to be listed
-#            here, separated by commas.  Any whitespace following the
-#            beginning of this list will be considered as a part of the
-#            variable name.
-#   	.i - List of variables used as input (remaining variables from .v
-#            list are constant).  .v must have been previously defined.
-#   	.o - List of variables used as output (remaining variables from .v
-#            list are garbage).  .v must have been previously defined.
-#   	.c - Values for constant input variables.  .v and .i must have been
-#            previously defined.
-#
-#   Once the four main tags have been defined, a line containing the text:
-#   "BEGIN" (without quotes, any case) signifies the beginning of the gate
-#   listing.
-#
-#   Gates:
-#
-#   	Gates should be written on a single line.  Gates will continue to
-#       be read until a line is reached that is not a proper gate, or end
-#       of file.
-#
-#	Format = Gx : 	G = 'T' or 'F' (lowercase acceptable) meaning
-#                           Toffoli or Fredkin gate
-#                 	x = Number of parameters
-#
-#	Parameters follow on the same line and, similar to the parameters
-#       for the tags above, are separated from the gate type by whitespace.
-#       Each parameter should be found in the .v list defined above, and
-#       should be separated by commas.
-#
-#   END statement terminates the input.
-
-.v a,b,c,d,e
-.i a,b,c
-.o d,c
-.c 0,1
-BEGIN
-t3 a,b,d
-t2 a,b
-t3 b,c,d
-t2 b,c
-END
