+++ test/ir/test_qasm3_parser.cpp
@@ -11,12 +11,13 @@
 #include "ir/QuantumComputation.hpp"
 #include "ir/operations/ClassicControlledOperation.hpp"
 #include "ir/operations/OpType.hpp"
-#include "ir/parsers/qasm3_parser/Exception.hpp"
-#include "ir/parsers/qasm3_parser/Parser.hpp"
-#include "ir/parsers/qasm3_parser/Scanner.hpp"
-#include "ir/parsers/qasm3_parser/Statement.hpp"
-#include "ir/parsers/qasm3_parser/Token.hpp"
-#include "ir/parsers/qasm3_parser/passes/ConstEvalPass.hpp"
+#include "qasm3/Exception.hpp"
+#include "qasm3/Importer.hpp"
+#include "qasm3/Parser.hpp"
+#include "qasm3/Scanner.hpp"
+#include "qasm3/Statement.hpp"
+#include "qasm3/Token.hpp"
+#include "qasm3/passes/ConstEvalPass.hpp"
 
 #include <cmath>
 #include <cstddef>
@@ -43,7 +44,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3) {
                                "\n"
                                "// lines */\n"
                                "bit[3] c;\n";
-  const auto qc = QuantumComputation::fromQASM(testfile);
+  const auto qc = qasm3::Importer::imports(testfile);
   EXPECT_EQ(qc.getNqubits(), 3);
   EXPECT_EQ(qc.getNcbits(), 3);
 }
@@ -53,7 +54,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3OldSyntax) {
                                "include \"stdgates.inc\";\n"
                                "qreg q[3];\n"
                                "creg r[3];\n";
-  const auto qc = QuantumComputation::fromQASM(testfile);
+  const auto qc = qasm3::Importer::imports(testfile);
   EXPECT_EQ(qc.getNqubits(), 3);
   EXPECT_EQ(qc.getNcbits(), 3);
 }
@@ -67,7 +68,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3GateDecl) {
                                "  x q2;\n"
                                "}\n"
                                "my_x q[0], q[1];\n";
-  auto qc = QuantumComputation::fromQASM(testfile);
+  auto qc = qasm3::Importer::imports(testfile);
 
   const std::string out = qc.toQASM();
   const std::string expected = "// i 0 1\n"
@@ -100,7 +101,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3CtrlModifier) {
       "ctrl @ swap q[0], q[1], q[2];\n"
       "ctrl @ rxx(pi) q[0], q[1], q[2];\n"
       "ctrl @ negctrl @ x q[0], q[1], q[2];\n";
-  auto qc = QuantumComputation::fromQASM(testfile);
+  auto qc = qasm3::Importer::imports(testfile);
 
   const std::string out = qc.toQASM();
   const std::string expected =
@@ -132,7 +133,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3InvModifier) {
                                "include \"stdgates.inc\";\n"
                                "qubit[1] q;\n"
                                "inv @ s q[0];\n";
-  auto qc = QuantumComputation::fromQASM(testfile);
+  auto qc = qasm3::Importer::imports(testfile);
 
   const std::string out = qc.toQASM();
   const std::string expected = "// i 0\n"
@@ -154,7 +155,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3CompoundGate) {
                                "  h q;\n"
                                "}\n"
                                "my_compound_gate q;";
-  auto qc = QuantumComputation::fromQASM(testfile);
+  auto qc = qasm3::Importer::imports(testfile);
 
   const std::string out = qc.toQASM();
   const std::string expected = "// i 0\n"
@@ -175,7 +176,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3ControlledCompoundGate) {
                                "  x q;\n"
                                "}\n"
                                "ctrl @ my_compound_gate q[0], q[1];\n";
-  auto qc = QuantumComputation::fromQASM(testfile);
+  auto qc = qasm3::Importer::imports(testfile);
 
   const std::string out = qc.toQASM();
   const std::string expected = "// i 0 1\n"
@@ -195,7 +196,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3ParamCompoundGate) {
                                "  rz(a) q;\n"
                                "}\n"
                                "my_compound_gate(1.0 * pi) q[0];\n";
-  auto qc = QuantumComputation::fromQASM(testfile);
+  auto qc = qasm3::Importer::imports(testfile);
 
   const std::string out = qc.toQASM();
   const std::string expected = "// i 0 1\n"
@@ -219,7 +220,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3Measure) {
                                "r2 = measure q;\n"
                                "measure q[1] -> r1;\n"
                                "measure q[1] -> r2[0];\n";
-  auto qc = QuantumComputation::fromQASM(testfile);
+  auto qc = qasm3::Importer::imports(testfile);
 
   const std::string out = qc.toQASM();
   const std::string expected = "// i 0 1\n"
@@ -245,7 +246,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3InitialLayout) {
                                "OPENQASM 3.0;\n"
                                "include \"stdgates.inc\";\n"
                                "qubit[2] q;\n";
-  auto qc = QuantumComputation::fromQASM(testfile);
+  auto qc = qasm3::Importer::imports(testfile);
 
   const std::string out = qc.toQASM();
   const std::string expected = "// i 1 0\n"
@@ -258,7 +259,6 @@ TEST_F(Qasm3ParserTest, ImportQasm3InitialLayout) {
 }
 
 TEST_F(Qasm3ParserTest, ImportQasm3ConstEval) {
-  std::stringstream ss{};
   const std::string testfile =
       "OPENQASM 3.0;\n"
       "include \"stdgates.inc\";\n"
@@ -266,10 +266,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3ConstEval) {
       "qubit[N * 2] q;\n"
       "ctrl @ x q[0], q[N * 2 - 1];\n"
       "x q;";
-
-  ss << testfile;
-  auto qc = QuantumComputation();
-  qc.import(ss, Format::OpenQASM3);
+  auto qc = qasm3::Importer::imports(testfile);
 
   const std::string out = qc.toQASM();
   const std::string expected = "// i 0 1 2 3 4 5 6 7\n"
@@ -299,7 +296,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3NonUnitary) {
                                "barrier q1, q2;\n"
                                "reset q1;\n"
                                "bit c = measure q1[0];\n";
-  auto qc = QuantumComputation::fromQASM(testfile);
+  auto qc = qasm3::Importer::imports(testfile);
 
   const std::string out = qc.toQASM();
   const std::string expected = "// i 0 1 2 3\n"
@@ -325,7 +322,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3IfStatement) {
                                "if (c == 1) {\n"
                                "  x q[1];\n"
                                "}";
-  auto qc = QuantumComputation::fromQASM(testfile);
+  auto qc = qasm3::Importer::imports(testfile);
 
   const std::string out = qc.toQASM();
   const std::string expected = "// i 0 1\n"
@@ -361,7 +358,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3IfElseStatement) {
                                  "  x q[0];\n"
                                  "  x q[1];\n"
                                  "}";
-    auto qc = QuantumComputation::fromQASM(testfile);
+    auto qc = qasm3::Importer::imports(testfile);
 
     const std::string out = qc.toQASM();
     const std::string expected =
@@ -397,7 +394,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3EmptyIfElse) {
                                "if (c == 1) {\n"
                                "} else {\n"
                                "}";
-  auto qc = QuantumComputation::fromQASM(testfile);
+  auto qc = qasm3::Importer::imports(testfile);
 
   const std::string out = qc.toQASM();
   const std::string expected = "// i 0 1\n"
@@ -415,7 +412,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3OutputPerm) {
   const std::string testfile = "// i 0 2 1 3\n"
                                "// o 3 0\n"
                                "qubit[4] q;\n";
-  const auto qc = QuantumComputation::fromQASM(testfile);
+  const auto qc = qasm3::Importer::imports(testfile);
 
   std::stringstream out{};
   QuantumComputation::printPermutation(qc.outputPermutation, out);
@@ -428,7 +425,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3OutputPerm) {
 TEST_F(Qasm3ParserTest, ImportQasm3OutputPermDefault) {
   const std::string testfile = "// i 0 2 1 3\n"
                                "qubit[4] q;\n";
-  const auto qc = QuantumComputation::fromQASM(testfile);
+  const auto qc = qasm3::Importer::imports(testfile);
 
   std::stringstream out{};
   QuantumComputation::printPermutation(qc.outputPermutation, out);
@@ -449,7 +446,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3IfElseNoBlock) {
                                "if (c == 1) {} else \n"
                                "  x q[1];\n"
                                "x q[0];\n";
-  auto qc = QuantumComputation::fromQASM(testfile);
+  auto qc = qasm3::Importer::imports(testfile);
 
   const std::string out = qc.toQASM();
   const std::string expected = "// i 0 1\n"
@@ -477,7 +474,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3InvalidStatementInBlock) {
                                "}";
   EXPECT_THROW(
       try {
-        const auto qc = QuantumComputation::fromQASM(testfile);
+        const auto qc = qasm3::Importer::imports(testfile);
       } catch (const qasm3::CompilerError& e) {
         EXPECT_EQ(e.message,
                   "Only quantum statements are supported in blocks.");
@@ -489,7 +486,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3InvalidStatementInBlock) {
 TEST_F(Qasm3ParserTest, ImportQasm3ImplicitInclude) {
   const std::string testfile = "qubit q;\n"
                                "h q[0];\n";
-  auto qc = QuantumComputation::fromQASM(testfile);
+  auto qc = qasm3::Importer::imports(testfile);
 
   const std::string out = qc.toQASM();
   const std::string expected = "// i 0\n"
@@ -506,7 +503,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3Qelib1) {
                                "include \"qelib1.inc\";\n"
                                "qubit q;\n"
                                "h q[0];\n";
-  auto qc = QuantumComputation::fromQASM(testfile);
+  auto qc = qasm3::Importer::imports(testfile);
 
   const std::string out = qc.toQASM();
   const std::string expected = "// i 0\n"
@@ -524,7 +521,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3Teleportation) {
                                "opaque teleport src, anc, tgt;\n"
                                "qubit[3] q;\n"
                                "teleport q[0], q[1], q[2];\n";
-  auto qc = QuantumComputation::fromQASM(testfile);
+  auto qc = qasm3::Importer::imports(testfile);
 
   const std::string out = qc.toQASM();
   const std::string expected =
@@ -546,7 +543,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3NestedGates) {
                                "gate my_x2 q1 { x q1; }\n"
                                "qubit[1] q;\n"
                                "my_x2 q[0];\n";
-  const auto qc = QuantumComputation::fromQASM(testfile);
+  const auto qc = qasm3::Importer::imports(testfile);
   EXPECT_EQ(qc.getNops(), 1);
   EXPECT_EQ(qc.at(0)->getType(), OpType::X);
 }
@@ -557,7 +554,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3AlternatingControl) {
                                "qubit[7] q;\n"
                                "ctrl @ negctrl(2) @ negctrl @ ctrl @ ctrl @ x "
                                "q[0], q[1], q[2], q[3], q[4], q[5], q[6];\n";
-  auto qc = QuantumComputation::fromQASM(testfile);
+  auto qc = qasm3::Importer::imports(testfile);
 
   const std::string out = qc.toQASM();
   const std::string expected = "// i 0 1 2 3 4 5 6\n"
@@ -576,7 +573,7 @@ TEST_F(Qasm3ParserTest, ImportQasmConstEval) {
                                "const uint N_1 = 0xa;\n"
                                "const uint N_2 = 8;\n"
                                "qubit[N_1 - N_2] q;\n";
-  auto qc = QuantumComputation::fromQASM(testfile);
+  auto qc = qasm3::Importer::imports(testfile);
 
   const std::string out = qc.toQASM();
   const std::string expected = "// i 0 1\n"
@@ -597,7 +594,7 @@ TEST_F(Qasm3ParserTest, ImportQasmBroadcasting) {
                                "h q1;\n"
                                "reset q2;\n"
                                "cx q1, q2;\n";
-  auto qc = QuantumComputation::fromQASM(testfile);
+  auto qc = qasm3::Importer::imports(testfile);
 
   const std::string out = qc.toQASM();
   const std::string expected = "// i 0 1 2 3\n"
@@ -627,7 +624,7 @@ TEST_F(Qasm3ParserTest, ImportQasmComparison) {
                                "if (c1 >= 0) { x q[0]; }\n"
                                "if (c1 == 0) { x q[0]; }\n"
                                "if (c1 != 0) { x q[0]; }\n";
-  auto qc = QuantumComputation::fromQASM(testfile);
+  auto qc = qasm3::Importer::imports(testfile);
 
   const std::string out = qc.toQASM();
   const std::string expected = "// i 0 1\n"
@@ -668,7 +665,7 @@ TEST_F(Qasm3ParserTest, ImportQasmNativeRedeclaration) {
                                "gate h q { U(pi/2, 0, pi) q; }\n"
                                "h q;\n"
                                "c1 = measure q;\n";
-  auto qc = QuantumComputation::fromQASM(testfile);
+  auto qc = qasm3::Importer::imports(testfile);
 
   const std::string out = qc.toQASM();
   const std::string expected = "// i 0\n"
@@ -689,7 +686,7 @@ TEST_F(Qasm3ParserTest, ImportQasm2CPrefix) {
                                "gate ccccx q1, q2, q3, q4, q5 {\n"
                                "}\n"
                                "ccccx q[0], q[1], q[2], q[3], q[4];\n";
-  auto qc = QuantumComputation::fromQASM(testfile);
+  auto qc = qasm3::Importer::imports(testfile);
 
   const std::string out = qc.toQASM();
   const std::string expected = "// i 0 1 2 3 4\n"
@@ -705,7 +702,7 @@ TEST_F(Qasm3ParserTest, ImportMCXGate) {
   const std::string testfile = "OPENQASM 3.0;\n"
                                "qubit[4] q;\n"
                                "mcx q[0], q[1], q[2], q[3];\n";
-  auto qc = QuantumComputation::fromQASM(testfile);
+  auto qc = qasm3::Importer::imports(testfile);
 
   const std::string out = qc.toQASM();
   const std::string expected = "// i 0 1 2 3\n"
@@ -740,7 +737,7 @@ TEST_F(Qasm3ParserTest, ImportMQTBenchCircuit) {
     measure eval[0] -> meas[0];
     measure q[0] -> meas[1];
   )";
-  auto qc = qc::QuantumComputation::fromQASM(qasm);
+  auto qc = qasm3::Importer::imports(qasm);
 
   const std::string out = qc.toQASM();
   const std::string expected = "// i 0 1\n"
@@ -775,7 +772,7 @@ TEST_F(Qasm3ParserTest, ImportMSGate) {
                                "ms(0.844396) q[0], q[1], q[2];"
                                "c = measure q;";
 
-  auto qc = QuantumComputation::fromQASM(testfile);
+  auto qc = qasm3::Importer::imports(testfile);
 
   const std::string out = qc.toQASM();
   const std::string expected = "// i 0 1 2\n"
@@ -798,7 +795,7 @@ TEST_F(Qasm3ParserTest, ImportQasm2CPrefixInvalidGate) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Usage of unknown gate 'cccck'.");
           throw;
@@ -814,7 +811,7 @@ TEST_F(Qasm3ParserTest, ImportQasm3CPrefix) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Usage of unknown gate 'ccccx'.");
           throw;
@@ -892,7 +889,7 @@ TEST_F(Qasm3ParserTest, ImportQasmParseOperators) {
                                "x **= 1;\n";
 
   ss << testfile;
-  qasm3::Parser parser(&ss, false);
+  qasm3::Parser parser(ss, false);
 
   const auto expectedTypes = std::vector{
       qasm3::AssignmentStatement::Type::PlusAssignment,
@@ -925,7 +922,7 @@ TEST_F(Qasm3ParserTest, ImportQasmParseUnaryExpressions) {
                                "sqrt(x)\n";
 
   ss << testfile;
-  qasm3::Parser parser(&ss, false);
+  qasm3::Parser parser(ss, false);
 
   const auto expectedTypes = std::vector{
       qasm3::UnaryExpression::Op::Sin, qasm3::UnaryExpression::Op::Cos,
@@ -953,7 +950,7 @@ TEST_F(Qasm3ParserTest, ImportQasmParseBinaryExpressions) {
                                "x > 5\n";
 
   ss << testfile;
-  qasm3::Parser parser(&ss, false);
+  qasm3::Parser parser(ss, false);
 
   const auto expectedTypes = std::vector{
       qasm3::BinaryExpression::Op::Power,
@@ -981,7 +978,7 @@ TEST_F(Qasm3ParserTest, ImportQasmUnknownQreg) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Usage of unknown quantum register.");
           throw;
@@ -998,7 +995,7 @@ TEST_F(Qasm3ParserTest, ImportQasmIndexOutOfBounds) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Index expression must be smaller than the "
                                "width of the quantum register.");
@@ -1017,7 +1014,7 @@ TEST_F(Qasm3ParserTest, ImportQasmIndexOutOfBoundsClassical) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Index expression must be smaller than the "
                                "width of the classical register.");
@@ -1035,7 +1032,7 @@ TEST_F(Qasm3ParserTest, ImportQasmDuplicateDeclaration) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Identifier 'q' already declared.");
           throw;
@@ -1052,7 +1049,7 @@ TEST_F(Qasm3ParserTest, ImportQasmInitConstRegWithMeasure) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Constant Evaluation: Constant declaration "
                                "initialization expression must be const.");
@@ -1070,7 +1067,7 @@ TEST_F(Qasm3ParserTest, ImportQasmAssignmentUnknownIdentifier) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Type Check Error: Type check failed.");
           throw;
@@ -1088,7 +1085,7 @@ TEST_F(Qasm3ParserTest, ImportQasmAssignmentConstVar) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Type Check Error: Type check failed.");
           throw;
@@ -1106,7 +1103,7 @@ TEST_F(Qasm3ParserTest, ImportQasmMultipleInputPermutations) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Multiple initial layout specifications found.");
           throw;
@@ -1124,7 +1121,7 @@ TEST_F(Qasm3ParserTest, ImportQasmMultipleOutputPermutations) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message,
                     "Multiple output permutation specifications found.");
@@ -1141,7 +1138,7 @@ TEST_F(Qasm3ParserTest, ImportQasmInvalidOpaqueGate) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Unsupported opaque gate 'asdf'.");
           throw;
@@ -1158,7 +1155,7 @@ TEST_F(Qasm3ParserTest, ImportQasmDuplicateGateDecl) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Gate 'my_x' already declared.");
           throw;
@@ -1174,7 +1171,7 @@ TEST_F(Qasm3ParserTest, ImportQasmDuplicateQubitArgGate) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Qubit 'q' already declared.");
           throw;
@@ -1191,7 +1188,7 @@ TEST_F(Qasm3ParserTest, ImportQasmUndeclaredGate) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Usage of unknown gate 'my_x'.");
           throw;
@@ -1209,7 +1206,7 @@ TEST_F(Qasm3ParserTest, ImportQasmInvalidGateTargets) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message,
                     "Gate 'my_x' takes 1 targets, but 2 were supplied.");
@@ -1227,7 +1224,7 @@ TEST_F(Qasm3ParserTest, ImportQasmInvalidGateControls) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message,
                     "Gate 'cx' takes 1 targets, but 0 were supplied.");
@@ -1245,7 +1242,7 @@ TEST_F(Qasm3ParserTest, ImportQasmInvalidGateModifiers) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message,
                     "Gate 'x' takes 2 controls, but only 1 were supplied.");
@@ -1264,7 +1261,7 @@ TEST_F(Qasm3ParserTest, ImportQasmGateCallNonConst) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Only const expressions are supported as gate "
                                "parameters, but found 'IdentifierExpr (c)'.");
@@ -1283,7 +1280,7 @@ TEST_F(Qasm3ParserTest, ImportQasmGateCallBroadcastingInvalidWidth) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(
               e.message,
@@ -1303,7 +1300,7 @@ TEST_F(Qasm3ParserTest, ImportQasmGateCallIndexingGateBody) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message,
                     "Gate arguments cannot be indexed within gate body.");
@@ -1321,7 +1318,7 @@ TEST_F(Qasm3ParserTest, ImportQasmGateMeasureInvalidSizes) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message,
                     "Classical and quantum register must have the same width "
@@ -1340,7 +1337,7 @@ TEST_F(Qasm3ParserTest, ImportQasmGateOldStyleDesignator) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "In OpenQASM 3.0, the designator has been "
                                "changed to `type[designator] identifier;`");
@@ -1357,7 +1354,7 @@ TEST_F(Qasm3ParserTest, ImportQasmGateExpectStatement) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Expected quantum statement, got '+'.");
           throw;
@@ -1372,7 +1369,7 @@ TEST_F(Qasm3ParserTest, ImportQasmGateVersionDeclaration) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(
               e.message,
@@ -1390,7 +1387,7 @@ TEST_F(Qasm3ParserTest, ImportQasmInvalidExpected) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Expected ',', got 'Identifier'.");
           throw;
@@ -1406,7 +1403,7 @@ TEST_F(Qasm3ParserTest, ImportQasmTypeMismatchAssignment) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Type Check Error: Type check failed.");
           throw;
@@ -1422,7 +1419,7 @@ TEST_F(Qasm3ParserTest, ImportQasmTypeMismatchBinaryExpr) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Type Check Error: Type check failed.");
           throw;
@@ -1437,7 +1434,7 @@ TEST_F(Qasm3ParserTest, ImportQasmConstNotInitialized) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message,
                     "Constant Evaluation: Constant declaration initialization "
@@ -1454,7 +1451,7 @@ TEST_F(Qasm3ParserTest, ImportQasmUnaryTypeMismatchLogicalNot) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Type Check Error: Type check failed.");
           throw;
@@ -1470,7 +1467,7 @@ TEST_F(Qasm3ParserTest, ImportQasmUnaryTypeMismatchBitwiseNot) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Type Check Error: Type check failed.");
           throw;
@@ -1485,7 +1482,7 @@ TEST_F(Qasm3ParserTest, ImportQasmBinaryTypeMismatch) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Type Check Error: Type check failed.");
           throw;
@@ -1501,7 +1498,7 @@ TEST_F(Qasm3ParserTest, ImportQasmAssignmentIndexType) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Type Check Error: Type check failed.");
           throw;
@@ -1516,7 +1513,7 @@ TEST_F(Qasm3ParserTest, ImportQasmUnknownIdentifier) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Type Check Error: Type check failed.");
           throw;
@@ -1531,7 +1528,7 @@ TEST_F(Qasm3ParserTest, ImportQasmUnknownQubit) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Type Check Error: Type check failed.");
           throw;
@@ -1546,7 +1543,7 @@ TEST_F(Qasm3ParserTest, ImportQasmNegativeTypeDesignator) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Type Check Error: Type check failed.");
           throw;
@@ -1565,7 +1562,7 @@ TEST_F(Qasm3ParserTest, ImportQasmRegisterDeclarationInDefinition) {
   EXPECT_THROW(
       {
         try {
-          const auto qc = QuantumComputation::fromQASM(testfile);
+          const auto qc = qasm3::Importer::imports(testfile);
         } catch (const qasm3::CompilerError& e) {
           EXPECT_EQ(e.message, "Expected quantum statement, got 'qubit'.");
           throw;
