+++ test/zx/CMakeLists.txt
@@ -8,4 +8,5 @@
 if(TARGET MQT::CoreZX)
   file(GLOB_RECURSE ZX_TEST_SOURCES *.cpp)
   package_add_test(mqt-core-zx-test MQT::CoreZX ${ZX_TEST_SOURCES})
+  target_link_libraries(mqt-core-zx-test PRIVATE MQT::CoreQASM)
 endif()
